Added ranking system for words and prototype guesser.
Found bug where if it's found that the solution has a double letter, the
solution list doesn't filter them correctly. Found bug where if a word
with a double letter is guessed and the answer doesn't have that double
letter, all solutions with 1 of that letter are also eliminated.

Possible optimisation by preferring guesses from the old answer list
instead of any random answer. Possible optimisation by adding the ranks
of the words as value to minimise instead of counting occurances to
maximise.