Fixed bug where filtering non dupes was being called incorrectly.
If guessed a word like sanes and the first s came back 1 and second 0,
the solver would have seen the dupe and the 1 and filtered non dupes.

Fixed by modifying dupes array to be added to after each character has
been checked.

NOTE, this introduces a new edge case where if sanes comes back 00002,
so the first s is grey and second is green, the solver will eliminate
all words with S's first and then eliminate all words without an S in
that position (every word).