Fixed (hopefully final) edge case in filter!
If sanes went 00002, the solver would first filter all words with the
letter S, then try to filter all words that didn't have an S in the
final position, leading to the empty list.

Solves this by adding all the 0 filters to a list and doing them after
the other filters, with a boolean check to make sure the letter isn't a
dupe.