Added a frequency table function in analyser.
This will be used to calculate the odds of a guessed character being in
the word, frequency is divided by the number of possible answers. It's
possible for a words frequency to be above 1 if the average number of
that letter in words is above 1 (ie there are a lot of possible
duplicates).