Finished the eval function.
Uses information theory to decide when the value gained from an
information is too low. If so, makes an explorationary guess.
Experimented with values of VALUE_OF_NEW_INFO constant, decided on 3 as
the magic number.

This has taken up so much time, I'd say it wasn't worth it if I didn't
love it.